Simplify diff
